Organise components used for index page into index folder
