+++ components/index/skills/index.jsx
