+++ lib/firebase.js
@@ -1,24 +1,22 @@
 import firebase from "firebase/app"
 import "firebase/firestore"
 
-export function loadDB() {
-    try {
-        var config = {
-            apiKey: process.env.FIREBASE_API_KEY,
-            authDomain: process.env.FIREBASE_AUTH_DOMAIN,
-            databaseURL: process.env.FIREBASE_DATABASE_URL,
-            projectId: process.env.FIREBASE_PROJECT_ID,
-            storageBucket: process.env.FIREBASE_STORAGE_BUCKET,
-            messagingSenderId: process.env.FIREBASE_MESSAGING_SENDER_ID,
-        }
-        firebase.initializeApp(config)
-    } catch (err) {
-        // we skip the "already exists" message which is
-        // not an actual error when we're hot-reloading
-        if (!/already exists/.test(err.message)) {
-            console.error("Firebase initialization error", err.stack)
-        }
+try {
+    var config = {
+        apiKey: process.env.FIREBASE_API_KEY,
+        authDomain: process.env.FIREBASE_AUTH_DOMAIN,
+        databaseURL: process.env.FIREBASE_DATABASE_URL,
+        projectId: process.env.FIREBASE_PROJECT_ID,
+        storageBucket: process.env.FIREBASE_STORAGE_BUCKET,
+        messagingSenderId: process.env.FIREBASE_MESSAGING_SENDER_ID,
     }
-
-    return firebase
+    firebase.initializeApp(config)
+} catch (err) {
+    if (!/already exists/.test(err.message)) {
+        console.error("Firebase initialization error", err.stack)
+    }
 }
+
+const projectDatabase = firebase.firestore()
+
+export { projectDatabase }
